# Make chunk_message more efficient