--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,3 @@
 .cache
 .coverage
+.hypothesis