--- a/framewirc/messages.py
+++ b/framewirc/messages.py
@@ -80,7 +80,7 @@ def _chunk_message(message, max_length):
     while lines:
         line = lines.popleft()
         line_bytes = line.encode()
-        # If the line fits, add it the the lines.
+        # If the line fits, add it to the accepted lines.
         if len(line_bytes) <= max_length:
             yield line_bytes
             continue