It was really slow before.

- `to_bytes` turns out to be pretty bad in a tight loop.
- We jumped straight into searching one char at a time before trying to split on space. (`rfind` is significantly faster.)

There's a little more that can be done here before merging. Mostly, I'm not a fan of looping over the chars one at a time. It's seriously inefficient. Most of the time, we're not going to find what we're looking for in the first few hundred iterations.

A binary search would be much better. I think we can also be more efficient about the minimum and maximum bounds based on the unicode restrictions, and the max length.

Note to future self: look into the [`bisect`](https://docs.python.org/3.6/library/bisect.html) module to see if it can abstract away rolling our own binary search.