@meshy 2017-04-21
I've been profiling this with a local script, but I'd be interested if anyone knows of a more formal way to write regression tests for the speed here.