@meshy 2018-01-29
Using this script:

```python
# profile.py
from framewirc.messages import chunk_message

a = '💩失' * 1200

for i in range(10001):
    chunk_message(a, 500)
```

And this command:

```bash
time python ./profile.py
```

This PR now changes the execution time from:

```
python ./profile.py  12.87s user 0.01s system 99% cpu 12.875 total
```

to:

```
python ./profile.py  2.01s user 0.00s system 99% cpu 2.013 total
```

Before @Ian-Foote's suggestion, I had got it down to:

```
python ./profile.py  8.54s user 0.01s system 99% cpu 8.550 total
```

---

I'd say there's still room for improvement, but this is pretty good start on premature optimisation `:)`