@meshy 2018-02-12
Well... I'm probably going to be fighting hypothesis on this for a little while... `:P`