@meshy 2018-02-12
@Ian-Foote, thanks to hypothesis, there are a couple of extra fixes in here.

- when splitting lines on spaces, the space is not removed from the end of the string.
- when a message fits exactly, it is not split.